requirements.txt
mpld3>=0.5.8
python-socketio>=5.7.2
flask >= 2.2.2
Flask-SocketIO >= 5.3.2
